Exceção customizada para o banco de dados - DabaBaseException (Falha de integridade)

+++ src/main/java/br/com/devsenior/dscommerce/services/ProductService.java
@@ -1,15 +1,19 @@
 package br.com.devsenior.dscommerce.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import br.com.devsenior.dscommerce.dto.ProductDTO;
 import br.com.devsenior.dscommerce.entities.Product;
 import br.com.devsenior.dscommerce.repositories.ProductRepository;
+import br.com.devsenior.dscommerce.services.exceptions.DataBaseException;
 import br.com.devsenior.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -45,16 +49,28 @@ public class ProductService {
 
     @Transactional // Sem o readOnly, pois irá salvar um dado no BD.
     public ProductDTO update(Long id, ProductDTO dto) {
-        Product entity = repository.getReferenceById(id);//Prepara o Objeto. Monitorado pela JPA.
-        copyDtoToEntity(dto, entity);
-        entity = repository.save(entity);
-        return new ProductDTO(entity);
-
+        try {
+            Product entity = repository.getReferenceById(id);//Prepara o Objeto. Monitorado pela JPA.
+            copyDtoToEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ProductDTO(entity);
+        } catch (EntityNotFoundException e) {
+            //Lançando minha exceção customizada
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        repository.deleteById(id);
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DataBaseException("Falha de integridade referencial");
+        }
+        
     }
 
     private void copyDtoToEntity(ProductDTO dto, Product entity) {
